Recover from and report uniqueness errors on load.
This was code I added to debug a problem with the task failing, down to
the fact that `get_or_create` is using the tuple of name, email and
Google ID to match people, which breaks if people change their name (or,
less likely, change their email).

The bug is fixable, but this seemed like useful code in its own right.